Move vec3.h to src/common/
