+++ src/TheRestOfYourLife/perlin.h
@@ -11,8 +11,8 @@
 // with this software. If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.
 //==================================================================================================
 
+#include "common/vec3.h"
 #include "random.h"
-#include "vec3.h"
 
 
 inline float perlin_interp(vec3 c[2][2][2], float u, float v, float w) {
